wp-dso-publish/requirements.txt
pyforms==4.0.3
pycryptodome==3.8.2
